+++ lib/src/model/event.g.dart
@@ -86,6 +86,8 @@ class _$Event extends Event {
   @override
   final DateTime? endR001;
   @override
+  final DateTime modifiedAt;
+  @override
   final String timezone;
   @override
   final bool isFree;
@@ -137,6 +139,7 @@ class _$Event extends Event {
       required this.startR001,
       required this.end,
       this.endR001,
+      required this.modifiedAt,
       required this.timezone,
       required this.isFree,
       required this.posterUrl,
@@ -177,6 +180,7 @@ class _$Event extends Event {
         startR001 == other.startR001 &&
         end == other.end &&
         endR001 == other.endR001 &&
+        modifiedAt == other.modifiedAt &&
         timezone == other.timezone &&
         isFree == other.isFree &&
         posterUrl == other.posterUrl &&
@@ -210,6 +214,7 @@ class _$Event extends Event {
     _$hash = $jc(_$hash, startR001.hashCode);
     _$hash = $jc(_$hash, end.hashCode);
     _$hash = $jc(_$hash, endR001.hashCode);
+    _$hash = $jc(_$hash, modifiedAt.hashCode);
     _$hash = $jc(_$hash, timezone.hashCode);
     _$hash = $jc(_$hash, isFree.hashCode);
     _$hash = $jc(_$hash, posterUrl.hashCode);
@@ -245,6 +250,7 @@ class _$Event extends Event {
           ..add('startR001', startR001)
           ..add('end', end)
           ..add('endR001', endR001)
+          ..add('modifiedAt', modifiedAt)
           ..add('timezone', timezone)
           ..add('isFree', isFree)
           ..add('posterUrl', posterUrl)
@@ -306,6 +312,10 @@ class EventBuilder implements Builder<Event, EventBuilder> {
   DateTime? _endR001;
   DateTime? get endR001 => _$this._endR001;
   set endR001(DateTime? endR001) => _$this._endR001 = endR001;
+
+  DateTime? _modifiedAt;
+  DateTime? get modifiedAt => _$this._modifiedAt;
+  set modifiedAt(DateTime? modifiedAt) => _$this._modifiedAt = modifiedAt;
 
   String? _timezone;
   String? get timezone => _$this._timezone;
@@ -401,6 +411,7 @@ class EventBuilder implements Builder<Event, EventBuilder> {
       _startR001 = $v.startR001;
       _end = $v.end;
       _endR001 = $v.endR001;
+      _modifiedAt = $v.modifiedAt;
       _timezone = $v.timezone;
       _isFree = $v.isFree;
       _posterUrl = $v.posterUrl;
@@ -455,6 +466,8 @@ class EventBuilder implements Builder<Event, EventBuilder> {
                 startR001, r'Event', 'startR001'),
             end: BuiltValueNullFieldError.checkNotNull(end, r'Event', 'end'),
             endR001: endR001,
+            modifiedAt: BuiltValueNullFieldError.checkNotNull(
+                modifiedAt, r'Event', 'modifiedAt'),
             timezone: BuiltValueNullFieldError.checkNotNull(
                 timezone, r'Event', 'timezone'),
             isFree: BuiltValueNullFieldError.checkNotNull(